fix:move session in nodejs + make controller for auth
