+++ src/hooks/use-session.ts
@@ -1,5 +1,4 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { supaClient } from "./supa-client.ts";
 
 export function useSupabaseSession() {
   const queryClient = useQueryClient();
@@ -11,12 +10,22 @@ export function useSupabaseSession() {
   } = useQuery({
     queryKey: ["session"],
     queryFn: async () => {
-      const { data, error } = await supaClient.auth.getSession();
-      if (error) {
+      try {
+        const response = await fetch("http://localhost:3001/auth/session", {
+          method: "GET",
+          credentials: "include",
+        });
+
+        if (!response.ok) {
+          throw new Error("Error fetching session");
+        }
+
+        const data = await response.json();
+        return data.session;
+      } catch (error) {
         console.error("Error fetching session:", error);
         return null;
       }
-      return data.session;
     },
     staleTime: 1000 * 60 * 5,
     cacheTime: 1000 * 60 * 10,
@@ -30,15 +39,26 @@ export function useSupabaseSession() {
       accessToken: string;
       refreshToken: string;
     }) => {
-      const { data, error } = await supaClient.auth.setSession({
-        access_token: accessToken,
-        refresh_token: refreshToken,
-      });
-      if (error) {
+      try {
+        const response = await fetch("http://localhost:3001/auth/session", {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          credentials: "include",
+          body: JSON.stringify({ accessToken, refreshToken }),
+        });
+
+        if (!response.ok) {
+          throw new Error("Error setting session");
+        }
+
+        const data = await response.json();
+        return data.session;
+      } catch (error) {
         console.error("Error setting session:", error);
         return null;
       }
-      return data.session;
     },
     onSuccess: (newSession) => {
       queryClient.setQueryData(["session"], newSession);
@@ -49,5 +69,5 @@ export function useSupabaseSession() {
     setSessionMutation.mutate({ accessToken, refreshToken });
   };
 
-  return { session, isLoading, setSession };
+  return { session, isLoading, setSession, error };
 }
