+++ src/server.ts
@@ -5,13 +5,12 @@ import session from "express-session";
 import passport from "passport";
 import { createServer } from "http";
 import { Server as SocketIOServer } from "socket.io";
-import authentificationRoutes from "./routes/authentificationRoutes.ts";
+import authRoutes from "./routes/authRoutes.ts";
 import serialPortRoutes from "./routes/serialPortRoutes.ts";
 import planRoutes from "./routes/planRoutes.ts";
 import wellnessNetworkRoutes from "./routes/wellnessNetworkRoutes.ts";
 import hmiRoutes from "./routes/hmiRoutes.ts";
 import userRoutes from "./routes/userRoutes.ts";
-import localServerRoutes from "./routes/localServerRoutes.ts";
 import wellnessNetworkRoutes from "./routes/wellnessNetworkRoutes.ts";
 import { getSerialPort } from "./managers/serialPort.ts";
 import "./config/passportConfig.ts";
@@ -28,7 +27,10 @@ const io = new SocketIOServer(httpServer, {
 });
 
 app.use(express.json());
-app.use(cors());
+app.use(cors({
+  origin: "http://localhost:1337",
+  credentials: true,
+}));
 
 // Session setup
 app.use(
@@ -43,7 +45,7 @@ app.use(
 app.use(passport.initialize());
 app.use(passport.session());
 
-app.use("/api", authentificationRoutes);
+app.use("/auth", authRoutes);
 app.use("/api", serialPortRoutes);
 app.use("/api", userRoutes);
 app.use("/api", planRoutes);
